Simplified logic and some perf improvements

+++ src/main/kotlin/com/ratelimiter/slot/Exceptions.kt
@@ -6,7 +6,7 @@ package com.ratelimiter.slot
  */
 class SlotAssignmentException(
     val eventId: String,
-    val windowsSearched: Int,
+    val windowsSearched: Long,
     message: String
 ) : RuntimeException(message)
